requirements.txt
beautifulsoup4==4.9.0
bokeh==2.0.2
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
idna==2.9
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
nltk==3.5
numpy==1.18.4
packaging==20.3
pandas==1.0.3
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
soupsieve==2.0
tornado==6.0.4
tqdm==4.46.0
typing-extensions==3.7.4.2
urllib3==1.25.9
vaderSentiment==3.3.1
wordcloud==1.7.0
